Fixed a lot of warnings.
svn path=/trunk/; revision=2012

+++ reactos/lib/msvcrt/io/read.c
@@ -10,13 +10,13 @@
 #include <crtdll/io.h>
 #include <windows.h>
 
-size_t	_read(int _fd, void *_buf, size_t _nbyte)
+size_t _read(int _fd, void *_buf, size_t _nbyte)
 {
-   size_t _rbyte;
+   DWORD _rbyte;
    
-   if (!ReadFile(_get_osfhandle(_fd),_buf,_nbyte,&_rbyte,NULL)) 
+   if (!ReadFile(_get_osfhandle(_fd),_buf,_nbyte,&_rbyte,NULL))
      {
 	return -1;
      }
-   return _rbyte;
+   return (size_t)_rbyte;
 }
